recipients_add: add/del recipient
the functions add/del/get_recipient(s) were updated.  If the original
reciepient accepts the message is not stored in module->msg->deletemyself .
The rest of the recipients are stored in module->msg->envrcpts .

Before the change envrcpts[0] was used for the original recipient
and envrcpts[1...] for the rest, but to implement this was error prone.